api-v2/services/common/provider-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0

motor==3.6.0
pymongo==4.9.2          # <-- compatível com motor 3.6.0

python-dotenv==1.0.1
redis==5.0.8
aiokafka==0.10.0
googlemaps==4.10.0
pydantic==1.10.15
python-jose==3.3.0

# Firebase dependencies
firebase-admin==6.4.0
google-auth==2.23.4